Checking what happens with Python 3.12
Hello and thanks for the great work on this! I develop an open-source software that depends on your work, and am currently blocked from supporting 3.12 because of fdasrsf. No pressure, since we have plenty of time to transition, but I thought that I would try this out and see what needs to be done in order to transition to 3.12 support.